check if the hologram is not null before deleting

+++ src/main/java/me/hsgamer/gamesinthebox/hologram/dh/DHHologram.java
@@ -29,6 +29,13 @@ public class DHHologram implements Hologram {
 
     @Override
     public void clear() {
-        hologram.delete();
+        try {
+            if (hologram != null) {
+                hologram.delete();
+                hologram = null;
+            }
+        } catch (Exception ignored) {
+            // IGNORED
+        }
     }
 }
